+++ Sources/SuperwallKit/Paywall/Presentation/PublicPresentation.swift
@@ -67,17 +67,17 @@ extension Superwall {
 
   // MARK: - Register
 
-  /// Registers an event to access a feature. When the event is added to a campaign on the Superwall dashboard, it can show a paywall.
+  /// Registers a placement to access a feature. When the placement is added to a campaign on the Superwall dashboard, it can show a paywall.
   ///
-  /// This shows a paywall to the user when: An event you provide is added to a campaign on the [Superwall Dashboard](https://superwall.com/dashboard); the user matches a rule in the campaign; and the user doesn't have an active subscription.
+  /// This shows a paywall to the user when: A placement you provide is added to a campaign on the [Superwall Dashboard](https://superwall.com/dashboard); the user matches an audience filter in the campaign; and the user doesn't have an active subscription.
   ///
-  /// Before using this method, you'll first need to create a campaign and add the event to the campaign on the [Superwall Dashboard](https://superwall.com/dashboard).
+  /// Before using this method, you'll first need to create a campaign and add the placement to the campaign on the [Superwall Dashboard](https://superwall.com/dashboard).
   ///
   /// The paywall shown to the user is determined by the rules defined in the campaign. When a user is assigned a paywall within a rule, they will continue to see that paywall unless you remove the paywall from the rule or reset assignments to the paywall.
   ///
   /// - Parameters:
-  ///   -  event: The name of the event you wish to register.
-  ///   - params: Optional parameters you'd like to pass with your event. These can be referenced within the rules of your campaign. Keys beginning with `$` are reserved for Superwall and will be dropped. Values can be any JSON encodable value, URLs or Dates. Arrays and dictionaries as values are not supported at this time, and will be dropped. Defaults to `nil`.
+  ///   -  placement: The name of the placement you wish to register.
+  ///   - params: Optional parameters you'd like to pass with your placement. These can be referenced within the audience filters of your campaign. Keys beginning with `$` are reserved for Superwall and will be dropped. Values can be any JSON encodable value, URLs or Dates. Arrays and dictionaries as values are not supported at this time, and will be dropped. Defaults to `nil`.
   ///   - handler: An optional handler whose functions provide status updates for a paywall. Defaults to `nil`.
   ///   - feature: A completion block containing a feature that you wish to paywall. Access to this block is remotely configurable via the [Superwall Dashboard](https://superwall.com/dashboard). If the paywall is set to _Non Gated_, this will be called when the paywall is dismissed or if the user is already paying. If the paywall is _Gated_, this will be called only if the user is already paying or if they begin paying. If no paywall is configured, this gets called immediately. This will not be called in the event of an error, which you can detect via the `handler`.
   public func register(
@@ -94,17 +94,17 @@ extension Superwall {
     )
   }
 
-  /// Registers an event which, when added to a campaign on the Superwall dashboard, can show a paywall.
+  /// Registers an placement which, when added to a campaign on the Superwall dashboard, can show a paywall.
   ///
-  /// This shows a paywall to the user when: An event you provide is added to a campaign on the [Superwall Dashboard](https://superwall.com/dashboard); the user matches a rule in the campaign; and the user doesn't have an active subscription.
+  /// This shows a paywall to the user when: A placement you provide is added to a campaign on the [Superwall Dashboard](https://superwall.com/dashboard); the user matches an audience filter in the campaign; and the user doesn't have an active subscription.
   ///
   /// Before using this method, you'll first need to create a campaign and add the event to the campaign on the [Superwall Dashboard](https://superwall.com/dashboard).
   ///
   /// The paywall shown to the user is determined by the rules defined in the campaign. When a user is assigned a paywall within a rule, they will continue to see that paywall unless you remove the paywall from the rule or reset assignments to the paywall.
   ///
   /// - Parameters:
-  ///   -  event: The name of the event you wish to register.
-  ///   - params: Optional parameters you'd like to pass with your event. These can be referenced within the rules of your campaign. Keys beginning with `$` are reserved for Superwall and will be dropped. Values can be any JSON encodable value, URLs or Dates. Arrays and dictionaries as values are not supported at this time, and will be dropped. Defaults to `nil`.
+  ///   -  placement: The name of the placement you wish to register.
+  ///   - params: Optional parameters you'd like to pass with your placement. These can be referenced within the audience filters of your campaign. Keys beginning with `$` are reserved for Superwall and will be dropped. Values can be any JSON encodable value, URLs or Dates. Arrays and dictionaries as values are not supported at this time, and will be dropped. Defaults to `nil`.
   ///   - handler: An optional handler whose functions provide status updates for a paywall. Defaults to `nil`.
   public func register(
     placement: String,
@@ -180,16 +180,16 @@ extension Superwall {
     }
   }
 
-  /// Objective-C-only convenience method. Registers an event which, when added to a campaign on the Superwall dashboard, can show a paywall.
+  /// Objective-C-only convenience method. Registers a placement which, when added to a campaign on the Superwall dashboard, can show a paywall.
   ///
-  /// This shows a paywall to the user when: An event you provide is added to a campaign on the [Superwall Dashboard](https://superwall.com/dashboard); the user matches a rule in the campaign; and the user doesn't have an active subscription.
+  /// This shows a paywall to the user when: A placement you provide is added to a campaign on the [Superwall Dashboard](https://superwall.com/dashboard); the user matches a rule in the campaign; and the user doesn't have an active subscription.
   ///
-  /// Before using this method, you'll first need to create a campaign and add the event to the campaign on the [Superwall Dashboard](https://superwall.com/dashboard).
+  /// Before using this method, you'll first need to create a campaign and add the placement to the campaign on the [Superwall Dashboard](https://superwall.com/dashboard).
   ///
-  /// The paywall shown to the user is determined by the rules defined in the campaign. When a user is assigned a paywall within a rule, they will continue to see that paywall unless you remove the paywall from the rule or reset assignments to the paywall.
+  /// The paywall shown to the user is determined by the audience filters defined in the campaign. When a user is assigned a paywall within an audience, they will continue to see that paywall unless you remove the paywall from the audience or reset assignments to the paywall.
   ///
   /// - Parameters:
-  ///   -  event: The name of the event you wish to register.
+  ///   -  placement: The name of the placement you wish to register.
   @available(swift, obsoleted: 1.0)
   @objc public func register(placement: String) {
     internallyRegister(placement: placement)